Splitting local dev
