+++ _config/site.yml
@@ -6,7 +6,7 @@ asciidoctor:
 slim|html:
   :format: :html
 slim:
-  :attr_list_delims: 
+  :attr_list_delims:
     '(': ')'
     '[': ']'
 download_manager_file_base_url: //developers.redhat.com/download-manager/file/
@@ -35,8 +35,8 @@ profile_base: &profile
   cdn_version: &cdn_version true
   log_faraday: true
   push_to_searchisko: true
-  minify: &minify true 
-  robots: 
+  minify: &minify true
+  robots:
     disallow: true
   deploy:
     <<: *deploy
@@ -75,7 +75,7 @@ profiles:
     base_url: http://<%= ENV['DRUPAL_HOST_IP'] %>:<%= ENV['DRUPAL_HOST_PORT'] %>/
     drupal_base_url: http://drupal/
 
-  development_cdn: 
+  development_cdn:
     <<: *development
     cdn_http_base: &dev_cdn_http_base http://localhost:4242/cdn
     cdn_out_dir: &dev_cdn_out_dir _site
@@ -100,7 +100,7 @@ profiles:
       path: /stg_htdocs/it-rhd-stg
 
   # Pull Request configuration
-  staging: 
+  staging:
     <<: *profile
     dcp_base_protocol_relative_url: //dcpbeta-searchisko.rhcloud.com/
     dcp_base_url: http://dcpbeta-searchisko.rhcloud.com/
@@ -112,7 +112,7 @@ profiles:
       path: /stg_htdocs/it-rhd-stg
 
   # Main staging configuration
-  staging_main: 
+  staging_main:
     <<: *profile
     base_url: http://developers.stage.redhat.com/
     download_manager_file_base_url: //developers.stage.redhat.com/download-manager/file/
@@ -129,7 +129,7 @@ profiles:
       path: /stg_htdocs/rhd-stg-main
 
   # Main staging configuration (on IT's Infr)
-  staging_main_it: 
+  staging_main_it:
     <<: *profile
     base_url: http://it-developers.stage.redhat.com/
     download_manager_file_base_url: //it-developers.stage.redhat.com/download-manager/file/
