+++ _docker/control.rb
@@ -11,60 +11,11 @@ require 'timeout'
 require 'erb'
 require 'resolv'
 require 'open3'
+require './lib/misc/Options.rb'
+require './lib/misc/FileHelpers.rb'
 
-class Options
-  def self.parse(args)
-    options = {:build => false, :restart => false, :drupal => false,
-               :stage_pr => false,
-               :awestruct => {:gen => false, :preview => false}}
 
-    opts_parse = OptionParser.new do |opts|
-      opts.banner = 'Usage: control.rb [options]'
-      opts.separator 'Specific options:'
-
-      docker_message = "Docker client connection info (i.e. tcp://example.com:1000). "\
-                       "DOCKER_HOST used if parameter not provided"
-      opts.on('-d', '--docker CONNECTION_INFO', String, docker_message) do |d|
-        options[:docker] = d
-      end
-
-      opts.on('-r', '--restart', 'Restart the containers') do |r|
-        options[:restart] = r
-      end
-
-      opts.on('-b', '--build', 'Build the containers') do |b|
-        options[:build] = b
-      end
-
-      opts.on('-g', '--generate', 'Run awestruct (clean gen)') do |r|
-        options[:awestruct][:gen] = true
-      end
-
-      opts.on('-p', '--preview', 'Run awestruct (clean preview)') do |r|
-        options[:awestruct][:preview] = true
-      end
-
-      opts.on('-u', '--drupal', 'Start up and enable drupal') do |u|
-        options[:drupal] = true
-      end
-
-      opts.on('--stage-pr PR_NUMBER', Integer, 'build for PR Staging') do |pr|
-        options[:stage_pr] = pr
-      end
-
-      # No argument, shows at tail.  This will print an options summary.
-      opts.on_tail('-h', '--help', 'Show this message') do
-        puts opts
-        exit
-      end
-    end
-
-    opts_parse.parse! args
-    options
-  end
-end
-
-def building_ci_job?()
+def building_ci_job?
   return ENV['BUILD_NUMBER']
 end
 
@@ -89,7 +40,11 @@ def modify_env(opts)
       abort "Unable to decrypt vault (#{e})"
     end
   end
+end
 
+def set_ports
+  #The environment names here are for INTERNAL purposes only and do not relate to
+  #variables generated by docker or docker-compose
   port_names = ['AWESTRUCT_HOST_PORT', 'DRUPAL_HOST_PORT', 'DRUPALMYSQL_HOST_PORT',
     'MYSQL_HOST_PORT', 'ES_HOST_PORT1', 'ES_HOST_PORT2', 'SEARCHISKO_HOST_PORT']
 
@@ -153,14 +108,7 @@ def block_wait_drupal_started
   end
 end
 
-def startup_services(opts)
-  puts 'Starting up services...'
-  if opts[:drupal]
-    execute_docker_compose :up, %w(-d elasticsearch mysql drupalmysql drupal searchisko searchiskoconfigure)
-  else
-    execute_docker_compose :up, %w(-d elasticsearch mysql searchisko searchiskoconfigure)
-  end
-
+def block_wait_searchisko_configure_started
   begin
     configure_service = Docker::Container.get('docker_searchiskoconfigure_1')
   rescue Excon::Errors::SocketError => se
@@ -179,44 +127,37 @@ def startup_services(opts)
     sleep 5
     configure_service = Docker::Container.get('docker_searchiskoconfigure_1')
   end
+end
+
+def startup_supporting_services(opts)
+  puts 'Starting up services...'
+  if opts[:drupal]
+    execute_docker_compose :up, %w(-d elasticsearch mysql drupalmysql drupal searchisko searchiskoconfigure)
+  else
+    execute_docker_compose :up, %w(-d elasticsearch mysql searchisko searchiskoconfigure)
+  end
+
+  block_wait_searchisko_configure_started
 
   # Check to see if Drupal is accepting connections before continuing
   block_wait_drupal_started if opts[:drupal]
 
-  if opts[:drupal]
-    execute_docker_compose :run, ['--no-deps', '--rm', '--service-ports', 'awestruct', 'rake git_setup clean preview[drupal]']
-  else
-    execute_docker_compose :run, ['--no-deps', '--rm', '--service-ports', 'awestruct', 'rake git_setup clean preview[docker]']
-  end
 end
 
 options = Options.parse ARGV
 
 puts Options.parse %w(-h) unless options_selected? options
 
-modify_env(options)
-
 #the docker url is taken from DOCKER_HOST env variable otherwise
 Docker.url = options[:docker] if options[:docker]
 
+modify_env(options)
+
 if options[:build] || options[:restart]
-# Output the new docker-compose file with the modified ports
+  set_ports()
+  # Output the new docker-compose file with the modified ports
   File.delete('docker-compose.yml') if File.exists?('docker-compose.yml')
   File.write('docker-compose.yml', ERB.new(File.read('docker-compose.yml.erb')).result)
-end
-
-def copy_if_changed(sourceFile, targetFile)
-  unless FileUtils.identical?(sourceFile, targetFile)
-    FileUtils.cp sourceFile, targetFile
-  end
-end
-
-def open_or_new(path)
-  if (File.exist?(path))
-    File.open(path)
-  else
-    File.new(path, 'w')
-  end
 end
 
 if options[:build]
@@ -225,12 +166,12 @@ if options[:build]
   parent_gemfile = File.open '../Gemfile'
   parent_gemlock = File.open '../Gemfile.lock'
 
-  target_gemfile = open_or_new(docker_dir + '/Gemfile')
-  target_gemlock = open_or_new(docker_dir + '/Gemfile.lock')
+  target_gemfile = FileHelpers.open_or_new(docker_dir + '/Gemfile')
+  target_gemlock = FileHelpers.open_or_new(docker_dir + '/Gemfile.lock')
 
   #Only copy if the file has changed. Otherwise docker won't cache optimally
-  copy_if_changed(parent_gemfile, target_gemfile)
-  copy_if_changed(parent_gemlock, target_gemlock)
+  FileHelpers.copy_if_changed(parent_gemfile, target_gemfile)
+  FileHelpers.copy_if_changed(parent_gemlock, target_gemlock)
 
   puts 'Building base docker image...'
   execute_docker(:build, '--tag=developer.redhat.com/base', './base')
@@ -244,7 +185,7 @@ end
 
 if options[:restart]
   execute_docker_compose :kill
-  startup_services(options)
+  startup_supporting_services(options)
 end
 
 if options[:awestruct][:gen]
@@ -252,7 +193,11 @@ if options[:awestruct][:gen]
 end
 
 if options[:awestruct][:preview]
-  execute_docker_compose :run, ['--no-deps', '--rm', '--service-ports', 'awestruct', 'rake clean preview[docker]']
+  if options[:drupal]
+    execute_docker_compose :run, ['--no-deps', '--rm', '--service-ports', 'awestruct', 'rake git_setup clean preview[drupal]']
+  else
+    execute_docker_compose :run, ['--no-deps', '--rm', '--service-ports', 'awestruct', 'rake git_setup clean preview[docker]']
+  end
 end
 
 if options[:stage_pr]
@@ -279,4 +224,3 @@ if options[:stage_pr]
   puts 'Searchisko configure started'
   execute_docker_compose :run, ['--no-deps', '--rm', '--service-ports', 'awestruct', "bundle exec rake create_pr_dirs[docker-pr,build,#{options[:stage_pr]}] clean deploy[staging_docker_pr]"]
 end
-
